cross-platform diagnostic
