+++ Library/Homebrew/extend/os/diagnostic.rb
@@ -1,2 +1,6 @@
 require "diagnostic"
-require "extend/os/mac/diagnostic" if OS.mac?
+if OS.mac?
+  require "extend/os/mac/diagnostic"
+elsif OS.linux?
+  require "extend/os/linux/diagnostic"
+end
